Did most of step 1, but there is an issue with the server flag in distributor.go
As stated in title, the first image is processed correctly, but once the second image is loaded we get a flag redefined error. That is probably because flags are defined in the distributor function, which gets called every time a test is run, but idk where else to define them.